更新完善

+++ FolderEny/EnyHelper.cs
@@ -1,4 +1,5 @@
-﻿using System;
+﻿using Microsoft.Win32;
+using System;
 using System.IO;
 using System.Runtime.InteropServices;
 using System.Security.AccessControl;
@@ -39,53 +40,126 @@ namespace FolderEny
             // 使用自动生成的密钥进行加密。
             return ASCIIEncoding.ASCII.GetString(desCrypto.Key);
         }
-
+        /// <summary>
+        /// 对字符串进行DES加密
+        /// </summary>
+        /// <param name="sourceString">待加密的字符串</param>
+        /// <returns>加密后的BASE64编码的字符串</returns>
+        public static string Encrypt(string sourceString)
+        {
+            byte[] btKey = Encoding.Default.GetBytes(sSecretKey);
+            byte[] btIV = Encoding.Default.GetBytes(sSecretKey);
+            DESCryptoServiceProvider des = new DESCryptoServiceProvider();
+            using (MemoryStream ms = new MemoryStream())
+            {
+                byte[] inData = Encoding.Default.GetBytes(sourceString);
+                try
+                {
+                    using (CryptoStream cs = new CryptoStream(ms, des.CreateEncryptor(btKey, btIV), CryptoStreamMode.Write))
+                    {
+                        cs.Write(inData, 0, inData.Length);
+                        cs.FlushFinalBlock();
+                    }
+                    return Convert.ToBase64String(ms.ToArray());
+                }
+                catch
+                {
+                    throw;
+                }
+            }
+        }
+        /// <summary>
+        /// 对DES加密后的字符串进行解密
+        /// </summary>
+        /// <param name="encryptedString">待解密的字符串</param>
+        /// <returns>解密后的字符串</returns>
+        public static string Decrypt(string encryptedString)
+        {
+            byte[] btKey = Encoding.Default.GetBytes(sSecretKey);
+            byte[] btIV = Encoding.Default.GetBytes(sSecretKey);
+            DESCryptoServiceProvider des = new DESCryptoServiceProvider();
+            using (MemoryStream ms = new MemoryStream())
+            {
+                byte[] inData = Convert.FromBase64String(encryptedString);
+                try
+                {
+                    using (CryptoStream cs = new CryptoStream(ms, des.CreateDecryptor(btKey, btIV), CryptoStreamMode.Write))
+                    {
+                        cs.Write(inData, 0, inData.Length);
+                        cs.FlushFinalBlock();
+                    }
+                    return Encoding.Default.GetString(ms.ToArray());
+                }
+                catch
+                {
+                    throw;
+                }
+            }
+        }
         /// <summary>
         /// 加密文件
         /// </summary>
-        /// <param name="sInputFilename">待加密的文件的完整路径</param>
-        /// <param name="sOutputFilename">加密后的文件的完整路径</param>
-        public static void EncryptFile(string sInputFilename, string sOutputFilename)
+        /// <param name="sourceFile">待加密的文件的完整路径</param>
+        /// <param name="destFile">加密后的文件的完整路径</param>
+        public static void EncryptFile(string sourceFile, string destFile)
         {
-            FileStream fsInput = new FileStream(sInputFilename, FileMode.Open, FileAccess.Read);
-
-            FileStream fsEncrypted = new FileStream(sOutputFilename, FileMode.Create, FileAccess.Write);
-            DESCryptoServiceProvider DES = new DESCryptoServiceProvider();
-            DES.Key = ASCIIEncoding.ASCII.GetBytes(sSecretKey);
-            DES.IV = ASCIIEncoding.ASCII.GetBytes(sSecretKey);
-            ICryptoTransform desencrypt = DES.CreateEncryptor();
-            CryptoStream cryptostream = new CryptoStream(fsEncrypted, desencrypt, CryptoStreamMode.Write);
-
-            byte[] bytearrayinput = new byte[fsInput.Length];
-            fsInput.Read(bytearrayinput, 0, bytearrayinput.Length);
-            cryptostream.Write(bytearrayinput, 0, bytearrayinput.Length);
-
-            cryptostream.Flush();
-            fsInput.Flush();
-            fsEncrypted.Flush();
-            cryptostream.Close();
-            fsInput.Close();
-            fsEncrypted.Close();
+            if (!File.Exists(sourceFile)) throw new FileNotFoundException("指定的文件路径不存在！", sourceFile);
+            byte[] btKey = Encoding.Default.GetBytes(sSecretKey);
+            byte[] btIV = Encoding.Default.GetBytes(sSecretKey);
+            DESCryptoServiceProvider des = new DESCryptoServiceProvider();
+            byte[] btFile = File.ReadAllBytes(sourceFile);
+            using (FileStream fs = new FileStream(destFile, FileMode.Create, FileAccess.Write))
+            {
+                try
+                {
+                    using (CryptoStream cs = new CryptoStream(fs, des.CreateEncryptor(btKey, btIV), CryptoStreamMode.Write))
+                    {
+                        cs.Write(btFile, 0, btFile.Length);
+                        cs.FlushFinalBlock();
+                    }
+                }
+                catch
+                {
+                    throw;
+                }
+                finally
+                {
+                    fs.Close();
+                }
+            }
         }
 
         /// <summary>
         /// 解密文件
         /// </summary>
-        /// <param name="sInputFilename">待解密的文件的完整路径</param>
-        /// <param name="sOutputFilename">解密后的文件的完整路径</param>
-        public static void DecryptFile(string sInputFilename, string sOutputFilename)
+        /// <param name="sourceFile">待解密的文件的完整路径</param>
+        /// <param name="destFile">解密后的文件的完整路径</param>
+        public static void DecryptFile(string sourceFile, string destFile)
         {
-            DESCryptoServiceProvider DES = new DESCryptoServiceProvider();
-            DES.Key = ASCIIEncoding.ASCII.GetBytes(sSecretKey);
-            DES.IV = ASCIIEncoding.ASCII.GetBytes(sSecretKey);
-
-            FileStream fsread = new FileStream(sInputFilename, FileMode.Open, FileAccess.Read);
-            ICryptoTransform desdecrypt = DES.CreateDecryptor();
-            CryptoStream cryptostreamDecr = new CryptoStream(fsread, desdecrypt, CryptoStreamMode.Read);
-            StreamWriter fsDecrypted = new StreamWriter(sOutputFilename);
-            fsDecrypted.Write(new StreamReader(cryptostreamDecr).ReadToEnd());
-            fsDecrypted.Flush();
-            fsDecrypted.Close();
+            if (!File.Exists(sourceFile)) throw new FileNotFoundException("指定的文件路径不存在！", sourceFile);
+            byte[] btKey = Encoding.Default.GetBytes(sSecretKey);
+            byte[] btIV = Encoding.Default.GetBytes(sSecretKey);
+            DESCryptoServiceProvider des = new DESCryptoServiceProvider();
+            byte[] btFile = File.ReadAllBytes(sourceFile);
+            using (FileStream fs = new FileStream(destFile, FileMode.Create, FileAccess.Write))
+            {
+                try
+                {
+                    using (CryptoStream cs = new CryptoStream(fs, des.CreateDecryptor(btKey, btIV), CryptoStreamMode.Write))
+                    {
+                        cs.Write(btFile, 0, btFile.Length);
+                        cs.FlushFinalBlock();
+                    }
+                }
+                catch
+                {
+                    throw;
+                }
+                finally
+                {
+                    fs.Close();
+                }
+            }
         }
         /// <summary>
         /// 加密文件夹
@@ -115,13 +189,28 @@ namespace FolderEny
                 destFolder = d.Parent.FullName + d.Name + enytype;
             d.MoveTo(destFolder);
             setFolder(destFolder, true);
+            //sethide();设置注册表隐藏
             return destFolder;
+        }
+        /// <summary>
+        /// 判断路径字符是否非法
+        /// </summary>
+        /// <param name="pathStr">表示路径的字符串</param>
+        /// <returns>true 为有效 false无效</returns>
+        public static bool isPathStr(string pathStr)
+        {
+            //Path.GetInvalidPathChars
+            //Path.GetInvalidFileNameChars
+            int index = pathStr.IndexOfAny(System.IO.Path.GetInvalidPathChars());
+            return index == -1;
         }
         /// <summary>
         /// 解密文件夹
         /// </summary>
         public static void DecryptFolder(string folderPath)
         {
+            if (!Directory.Exists(folderPath))
+                return;
             DirectoryInfo d = new DirectoryInfo(folderPath);
             File.Delete(folderPath + "\\p.xml");
             string destFolder = folderPath.Substring(0, folderPath.LastIndexOf("."));
@@ -221,6 +310,32 @@ namespace FolderEny
             //fileAcl.RemoveAccessRuleSpecific(everyoneRule);
             //从当前文件或目录移除单个匹配的允许或拒绝访问控制列表 (ACL) 权限。
         }
+        /// <summary>
+        /// 设置WINDOWS不能查看隐藏文件和系统文件
+        /// </summary>
+        public static void sethide()
+        {
+            //ClassesRoot -->注册表基项HKEY_CLASSES_ROOT --> 包含了所有应用程序运行时必需的信息
+            //CurrentUser -->注册表基项 HKEY_CURRENT_USER -->管理系统当前的用户信息
+            //Users -->注册表基项HKEY_USERS -->仅包含了缺省用户设置和登录用户的信息
+            //LocalMachine -->注册表基项HKEY_LOCAL_MACHINE -->保存了注册表里的所有与这台计算机有关的配置信息
+            //CurrentConfig -->注册表基项 HKEY_CURRENT_CONFIG -->允许软件和设备驱动程序员很方便的更新注册表，而不涉及到多个配置文件信息
+            //1表示选中 2的意思是随着另一个的选中而变化 把NOHIDDEN的CheckedValue值改成2，把SHOWALL的CheckedValue 值改成0，这样不管以后在面板里怎么调，始终都不会显示隐藏的文件和文件夹了
+            RegistryKey hklm = Microsoft.Win32.Registry.LocalMachine;
+            RegistryKey rkvalue = hklm.OpenSubKey(@"SOFTWARE\Microsoft\Windows\CurrentVersion\Explorer\Advanced\Folder\Hidden\SHOWALL", true);
+            rkvalue.SetValue("CheckedValue", 0);
+            rkvalue.SetValue("DefaultValue", 0);
+            rkvalue = hklm.OpenSubKey(@"SOFTWARE\Microsoft\Windows\CurrentVersion\Explorer\Advanced\Folder\Hidden\NOHIDDEN", true);
+            rkvalue.SetValue("CheckedValue", 0);
+            rkvalue.SetValue("DefaultValue", 0);
+            rkvalue = hklm.OpenSubKey(@"SOFTWARE\Microsoft\Windows\CurrentVersion\Explorer\Advanced\Folder\SuperHidden", true);
+            rkvalue.SetValue("CheckedValue", 1);
+            rkvalue.SetValue("DefaultValue", 1);
+            RegistryKey hkcu = Microsoft.Win32.Registry.CurrentUser;
+            RegistryKey rkv = hkcu.OpenSubKey(@"Software\Microsoft\Windows\CurrentVersion\Explorer\Advanced", true);
+            rkv.SetValue("Hidden", 0);
+            rkv.SetValue("ShowSuperHidden", 1);
 
+        }
     }
 }
